fix(vscode): replace mocha with inline test harness to drop RCE vuln
mocha@10 pulls serialize-javascript <=7.0.2 (GHSA-5c6j-r48x-rmvq, high).
Mocha was only added to satisfy @vscode/test-electron's requirement that
extensionTestsPath export run(): Promise<void>.

Replace it with a zero-dependency inline harness in index.ts: a typed
Test array, sequential await-based execution, and a rejection on any
failure. The tests from ask-roundtrip.test.ts are inlined; that file is
deleted. Remove mocha + @types/mocha from devDependencies and "mocha"
from tsconfig.integration.json types.
