+++ packages/vscode-extension/tsconfig.integration.json
@@ -4,7 +4,7 @@
     "module": "CommonJS",
     "moduleResolution": "node",
     "ignoreDeprecations": "6.0",
-    "types": ["node", "mocha"],
+    "types": ["node"],
     "rootDir": ".",
     "noEmit": false,
     "outDir": "test-out"
